reimplement color term persistence as Config instead of global
A global being set in a config override file does not pickle, since
it's relying on a side-effect.  Therefore, in order for color term
selection to be picklable, it needs to be in a Config.  This new
design uses two layers of ConfigDictField to create a menu for
selecting the astrometry_net_data version and target band of
interest.  The bottom-level ColortermConfig can construct a
Colorterm instance.

Because the full list of colorterms is no longer globally accessible,
the Astrometry needs to be told whether to pull all available fluxes
(it can no longer dig into some other module's private information);
it's best to default this to 'True' to ensure the user gets what he
needs.