Add wrapper for astrometry.net catalog to cache loading
Loading the entire catalog currently requires reading all the index files,
just to get the metadata (the healpix and nside), which is slow.
Here, the multiindex_t and list of multiindex_t are wrapped by the new
MultiIndexCache and AstrometryNetCatalog classes, respectively.  An
AstrometryNetCatalog can have a cache written for faster loading of the
metadata.  The cache can be generated with a new bin script, andCache.py.

The Astrometry class really needs to be cleaned up more, and more could
be moved into the AstrometryNetCatalog class, but I am trying to avoid
the consequent API changes for now, as this will soon be redone on the
LSST side.