MultiIndexCache: fix accounting of loaded state in fromFilenameList
MultiIndexCache.fromFilenameList was calling "read", which does not
set the "_loaded" member.  Subsequent calls to "reload" would
therefore attempt to load the data through astrometry.net, which
uses assert() to test that it has not been loaded.

Instead of calling "read", we should call "reload", which does the
accounting (calls "read" and sets "_loaded").